requirements.txt
bertopic==0.9.4
searchtweets-v2==1.1.1
pandas==1.4.2
black==22.3.0
numba==0.55.1
nltk==3.7
streamlit==1.8.1